refactor: ロビーの詳細画面をpages layerに移動
