+++ src/app/lobby/create/page.tsx
@@ -1,6 +1,6 @@
-import { NavigationBar } from "@/features/navigation-bar/components/navigation-bar";
 import LobbyForm from "@/pages/lobby-creation/ui/lobby-form";
 import { Card } from "@/shared/ui/card";
+import { NavigationBar } from "@/widgets/navigation-bar/ui/navigation-bar";
 
 export default async function Page({
 	searchParams,
