+++ src/pages/lobby-detail/api/get-lobbies.ts
