+++ src/pages/lobby-detail/config/school-time.json
