+++ src/pages/lobby-detail/model/use-chat.tsx
@@ -2,8 +2,8 @@
 
 import { type FormEvent, useEffect, useRef, useState } from "react";
 import { useAuth } from "@/features/auth/auth-provider";
-import { useLobby } from "../providers/lobby-provider";
-import type { Chat } from "../types";
+import type { Chat } from "../../../features/lobby/types";
+import { useLobby } from "./lobby-provider";
 
 const CHAT_COOLTIME_SECONDS = 300; // 5分
