+++ src/pages/lobby-detail/model/use-lobby-subscription.tsx
@@ -4,9 +4,9 @@ import type { RealtimeChannel } from "@supabase/supabase-js";
 import { useEffect, useState } from "react";
 import { useAuth } from "@/features/auth/auth-provider";
 import { createClient } from "@/shared/api/supabase/client";
+import type { Member } from "../../../features/lobby/types";
 import { joinLobby } from "../actions/join-lobby";
 import { leaveLobby } from "../actions/leave-lobby";
-import type { Member } from "../types";
 
 export function useLobbySubscription(lobbyId: string, goal: string | null) {
 	const { user } = useAuth();
