+++ src/pages/lobby-detail/ui/lobby-detail.tsx
