+++ src/pages/lobby-detail/ui/lobby-tips.tsx
