+++ src/pages/lobby-list/ui/lobby-card.tsx
@@ -8,7 +8,7 @@ import {
 } from "lucide-react";
 import Link from "next/link";
 import type { Lobby } from "@/features/lobby/types";
-import { useTimer } from "@/features/timer/hooks/use-timer";
+import { useTimer } from "@/pages/lobby-detail/model/use-timer";
 import { GetRelativeTime } from "@/shared/lib/get-relative-time";
 import { Button } from "@/shared/ui/button";
 import { Card, CardContent, CardFooter, CardHeader } from "@/shared/ui/card";
