+++ src/widgets/navigation-bar/ui/login-button.tsx
