+++ src/widgets/navigation-bar/ui/navigation-bar.tsx
